requirements.txt
torch
transformers
pytesseract
opencv-python
Pillow
flask